Add files from momlevel v1.0.2
This commit introduces files from momlevel v1.0.2 that will be used for
steric calculations with MPAS-Ocean.  momlevel is developed at
https://github.com/jkrasting/momlevel

The license for momlevel is:
Software code created by U.S. Government employees is not subject to copyright in the United States (17 U.S.C. §105).
The United States/Department of Commerce reserve all rights to seek and obtain copyright protection in countries
otherthan the United States for Software authored in its entirety by the Department of Commerce. To this end, the
Department of Commerce hereby grants to Recipient a royalty-free, nonexclusive license to use, copy, and create
derivative works of the Software outside of the United States.”

Future commits will adapt this code for use with MPAS-Ocean and
MPAS-Analysis as a derivative work.