Remove unneeded functions from momlevel and adjust imports
Removal of unneeded functions was for simplicity but also to eliminate
dependencies for python packages not currently in mpas-analysis.  The
functions using them are not relevant to the stric analysis.

This commit also adjusts imports for the modules from momlevel to make
them part of mpas-analysis.

With these changes, my standalone script can use mpas-analysis imports
to run successfully.