Add time_series_steric_anomaly analysis
This commit adds the basic global steric analysis from my standalone
script using momlevel into a new mpas-analysis analysis module.  None
of the existing base classes seemed to be a good fit for interfacing
with the momlevel code added to ocean/sealevel, so there are a lot of
low-level operations in the new module.

This first steric analysis simply plots the anomaly in the global mean
steric height as a time-series.